+++ src/crypto/internal/edwards25519/doc.go
