+++ src/crypto/internal/edwards25519/field/_asm/go.mod
