+++ src/crypto/internal/edwards25519/field/fe_amd64.go
