+++ src/crypto/internal/edwards25519/scalar.go
