+++ src/crypto/internal/edwards25519/scalarmult.go
