+++ src/crypto/internal/edwards25519/tables.go
