+++ src/crypto/internal/edwards25519/tables_test.go
